[FEATURE glimmer-custom-component-manager] Custom Component Manager
This is the first step towards opening up new APIs for Ember.js community as per
[Custom Components RFC](https://github.com/emberjs/rfcs/blob/custom-components/text/0000-custom-components.md).

This will allow the following:

+ add-on authors can now start experimenting with Glimmer API while continue to support existing Ember.js components
+ developers will be able to _opt into_ using custom component managers

This change should be considered cutting-edge and introduced behind a feature flag `GLIMMER_CUSTOM_COMPONENT_MANAGER`.

To enable this feature flag in `config/environment.js` of your Ember application:

```js
module.exports = function(environment) {
  let ENV = {
    ...
    EmberENV: {
      FEATURES: {
        'glimmer-custom-component-manager': true
      }
    }
  };

  ...

  return ENV;
};
```

Or in Glimmer application, `config/environment.js`:

```js
module.exports = function(environment) {
  let ENV = {
    ...

    GlimmerENV: {
      FEATURES: {
        'glimmer-custom-component-manager': true
      }
    }
  };

  return ENV;
};
```

After that, you should be able to opt into registering your own component managers:

+ for applications

App with classic structure:

```
my-app/
  component-managers/
    nyan.js
  components/
  templates/
    components/
      nyan-cat.hbs
```

```hbs
{{!- tempates/components/nyan-cat.hbs -}}
{{use-component-manager "nyan"}}

<h1>🐱</h1>
```

App with Module Unification:

```
my-app/
  src/
    ui/
      component-managers/
        nyan.js
      components/
        nyan-cat/
          template.hbs
```

```hbs
{{!- src/ui/components/nyan-cat.hbs -}}
{{use-component-manager "nyan"}}

<h1>🐱</h1>
```

+ for add-ons

```
nyan-cat-addon/
  addon/
    component-managers/
      glimmer.js
  app/
    templates/
      components/
        nyan-cat.hbs
```
