+++ features.json
@@ -6,7 +6,8 @@
     "ember-metal-weakmap": null,
     "ember-glimmer-allow-backtracking-rerender": null,
     "ember-routing-router-service": null,
-    "ember-engines-mount-params": null
+    "ember-engines-mount-params": null,
+    "glimmer-custom-component-manager": null
   },
   "deprecations": {
     "container-lookupFactory": "2.12.0",
